fix: address Sourcery review on PR #306
- Extract findSystemPython/venvPython/venvIsWorking into a shared
  scripts/mcpVenvProbe.mjs, consumed by both postinstall-mcp-setup.mjs
  and ensureMcpSetup.ts, instead of duplicating the logic (kept as
  plain ESM so the postinstall script needs zero build step).
- Document why the flagged spawn/spawnSync calls aren't a command-
  injection risk: cmd/python args are always a fixed candidate or a
  path constructed from a known .venv dir, never user input, and none
  of the calls use a shell.
- Wrap the PowerShell $ErrorActionPreference override in try/finally
  so it can't leak "Continue" for the rest of the script on an
  unexpected exception.
