+++ scripts/postinstall-mcp-setup.mjs
@@ -15,6 +15,7 @@ import { existsSync } from "node:fs";
 import { spawnSync } from "node:child_process";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
+import { venvPython, venvIsWorking, findSystemPython } from "./mcpVenvProbe.mjs";
 
 const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
 const mcpDir = path.join(root, "olive-mcp-server");
@@ -38,39 +39,10 @@ if (!existsSync(mcpDir)) {
   process.exit(0);
 }
 
-function venvPython() {
-  const winPy = path.join(mcpDir, ".venv", "Scripts", "python.exe");
-  const nixPy = path.join(mcpDir, ".venv", "bin", "python");
-  if (existsSync(winPy)) return winPy;
-  if (existsSync(nixPy)) return nixPy;
-  return null;
-}
-
-function venvIsWorking(python) {
-  // `mcp` (pinned <2) has no __version__ attribute -- just check it imports.
-  const r = spawnSync(python, ["-c", "import mcp"], {
-    encoding: "utf8",
-    env: { ...process.env, PYTHONPATH: mcpDir },
-  });
-  return r.status === 0;
-}
-
-const existingPython = venvPython();
-if (existingPython && venvIsWorking(existingPython)) {
+const existingPython = venvPython(mcpDir);
+if (existingPython && venvIsWorking(existingPython, mcpDir)) {
   log("MCP server venv already set up, nothing to do.");
   process.exit(0);
-}
-
-function findSystemPython() {
-  const candidates = process.platform === "win32" ? ["python", "python3"] : ["python3", "python"];
-  for (const cmd of candidates) {
-    const r = spawnSync(cmd, ["--version"], { encoding: "utf8" });
-    if (r.status === 0) {
-      const match = /Python 3\.(\d+)/.exec(r.stdout || r.stderr || "");
-      if (match && Number(match[1]) >= 10) return cmd;
-    }
-  }
-  return null;
 }
 
 if (!findSystemPython()) {
